refactor canvas and scene components, rename r3f dir
