+++ components/r3f/canvas.tsx
@@ -4,9 +4,9 @@ import dynamic from "next/dynamic";
 import { Canvas } from "@react-three/fiber";
 import { OrbitControls } from "@react-three/drei";
 import { SceneProvider } from "@/contexts/scene-context";
-import SceneControls from "@/components/rtf/scene-controls";
+import SceneControls from "@/components/r3f/scene-controls";
 
-const Scene = dynamic(() => import("@/components/rtf/scene"), { ssr: false });
+const Scene = dynamic(() => import("@/components/r3f/scene"), { ssr: false });
 
 export default function ThreeCanvas() {
   return (
