+++ components/r3f/scene-controls.tsx
