+++ components/r3f/scenes/torus-scene.tsx
@@ -1,0 +1,15 @@
+"use client";
+
+import TorusMesh from "@/components/r3f/meshes/torus-mesh";
+
+type TorusSceneProps = {
+  primaryColor: string;
+};
+
+export default function TorusScene({ primaryColor }: TorusSceneProps) {
+  return (
+    <>
+      <TorusMesh color={primaryColor} />
+    </>
+  );
+}
